requirements.txt
# Core scientific stack
numpy>=1.23
pandas>=2.0
xarray>=2023.12.0
netcdf4>=1.6
scipy>=1.10
PyYAML>=6.0.2

# Plotting
matplotlib>=3.7